app/requirements.txt
beautifulsoup4==4.9.3
selenium==3.141.0
requests==2.23.0
python-telegram-bot==13.1